+++ src/main/java/DBConnection/IConnection.java
@@ -1,0 +1,5 @@
+package DBConnection;
+
+public interface IConnection {
+     void sqliteConnect( String sqlUrl,String PackageNameForModels );
+}
