--- src/main/java/databaseBank/connectionInterface.java
@@ -1,7 +1,0 @@
-package databaseBank;
-
-import java.sql.Connection;
-
-public interface connectionInterface {
-     void sqliteConnect( String sqlUrl,String PackageNameForModels );
-}
