--- src/main/java/services/createPkg/createAnnotations/makeTable.java
@@ -1,11 +1,0 @@
-package services.createPkg.createAnnotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface makeTable {
-}
