--- src/main/java/services/createPkg/createAnnotations/sqliteColumn.java
@@ -1,18 +1,0 @@
-package services.createPkg.createAnnotations;
-
-
-import services.createPkg.createAnnotations.constraintList.constraintType;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface sqliteColumn {
-   double DEFAULT() default (0.000);
-   String CHECK() default ("SQL Condition");
-   constraintType constraint1() default constraintType.empty;
-   constraintType constraint2() default constraintType.empty;
-}
