--- src/main/java/services/createPkg/interFace/interFace.java
@@ -1,7 +1,0 @@
-package services.createPkg.interFace;
-
-public interface interFace<T> {
-
-    int newTable(T object) throws Exception;
-
-}
