+++ src/main/java/services/creates/acreate/MakeTable.java
@@ -1,0 +1,11 @@
+package services.creates.acreate;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface MakeTable {
+}
