+++ src/main/java/services/creates/icreate/ICreate.java
@@ -1,0 +1,7 @@
+package services.creates.icreate;
+
+public interface ICreate<T> {
+
+    int newTable(T object) throws Exception;
+
+}
