--- src/main/java/services/deletePkg/delete/delete.java
@@ -1,6 +1,0 @@
-package services.deletePkg.delete;
-
-import services.deletePkg.interFace.interFace;
-
-public class delete implements interFace {
-}
