--- src/main/java/services/deletePkg/interFace/interFace.java
@@ -1,4 +1,0 @@
-package services.deletePkg.interFace;
-
-public interface interFace {
-}
