+++ src/main/java/services/deletes/delete/DeleteWorker.java
@@ -1,0 +1,6 @@
+package services.deletes.delete;
+
+import services.deletes.idelete.IDelete;
+
+public class DeleteWorker implements IDelete {
+}
