+++ src/main/java/services/deletes/idelete/IDelete.java
@@ -1,0 +1,4 @@
+package services.deletes.idelete;
+
+public interface IDelete {
+}
