+++ src/main/java/services/fetchs/fetch/FetchWorker.java
@@ -1,0 +1,8 @@
+package services.fetchs.fetch;
+
+import services.fetchs.ifetch.IFetch;
+
+public class FetchWorker implements IFetch {
+
+
+}
