+++ src/main/java/services/fetchs/ifetch/IFetch.java
@@ -1,0 +1,4 @@
+package services.fetchs.ifetch;
+
+public interface IFetch {
+}
