--- src/main/java/services/insertPkg/interFace/interFace.java
@@ -1,8 +1,0 @@
-package services.insertPkg.interFace;
-
-import java.util.List;
-
-public interface interFace<T> {
-     int insertRow(T object);
-     int insertMultiRow(List<T> object);
-}
