+++ src/main/java/services/inserts/iinsert/IInsert.java
@@ -1,0 +1,8 @@
+package services.inserts.iinsert;
+
+import java.util.List;
+
+public interface IInsert<T> {
+     int insertRow(T object);
+     int insertMultiRow(List<T> object);
+}
