--- src/main/java/services/returnPkg/AReturn/AReturn.java
@@ -1,8 +1,0 @@
-package services.returnPkg.AReturn;
-
-import services.returnPkg.interFace.interFace;
-
-public class AReturn implements interFace {
-
-
-}
