--- src/main/java/services/returnPkg/interFace/interFace.java
@@ -1,4 +1,0 @@
-package services.returnPkg.interFace;
-
-public interface interFace {
-}
